requirements-develop.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements-develop.txt requirements-develop.in
#
asgiref==3.6.0
    # via
    #   -r requirements.txt
    #   django
attrs==22.2.0
    # via
    #   -r requirements.txt
    #   jsonschema
black==22.12.0
    # via -r requirements-develop.in
certifi==2022.12.7
    # via
    #   -r requirements.txt
    #   requests
charset-normalizer==3.0.1
    # via
    #   -r requirements.txt
    #   requests
click==8.1.3
    # via black
coreapi==2.3.3
    # via
    #   -r requirements.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements.txt
    #   coreapi
    #   drf-yasg
coverage==7.1.0
    # via -r requirements-develop.in
django==4.1.5
    # via
    #   -r requirements.txt
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-yasg
django-autoslug==1.9.8
    # via -r requirements.txt
django-debug-toolbar==3.8.1
    # via -r requirements-develop.in
django-extensions==3.2.1
    # via -r requirements.txt
django-filter==22.1
    # via -r requirements.txt
django-widget-tweaks==1.4.12
    # via -r requirements.txt
djangorestframework==3.14.0
    # via
    #   -r requirements.txt
    #   drf-yasg
drf-yasg[validation]==1.21.4
    # via -r requirements.txt
factory-boy==3.2.1
    # via -r requirements-develop.in
faker==16.6.1
    # via factory-boy
flake8==6.0.0
    # via -r requirements-develop.in
idna==3.4
    # via
    #   -r requirements.txt
    #   requests
importlib-metadata==6.0.0
    # via
    #   -r requirements.txt
    #   markdown
inflection==0.5.1
    # via
    #   -r requirements.txt
    #   drf-yasg
isort==5.12.0
    # via -r requirements-develop.in
itypes==1.2.0
    # via
    #   -r requirements.txt
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements.txt
    #   coreschema
jsonschema==4.17.3
    # via
    #   -r requirements.txt
    #   swagger-spec-validator
markdown==3.4.1
    # via -r requirements.txt
markupsafe==2.1.2
    # via
    #   -r requirements.txt
    #   jinja2
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==23.0
    # via
    #   -r requirements.txt
    #   drf-yasg
pathspec==0.11.0
    # via black
pillow==9.4.0
    # via -r requirements.txt
platformdirs==2.6.2
    # via black
psycopg2-binary==2.9.5
    # via -r requirements.txt
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyrsistent==0.19.3
    # via
    #   -r requirements.txt
    #   jsonschema
python-dateutil==2.8.2
    # via faker
pytz==2022.7.1
    # via
    #   -r requirements.txt
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via
    #   -r requirements.txt
    #   swagger-spec-validator
raven==6.10.0
    # via -r requirements.txt
requests==2.28.2
    # via
    #   -r requirements.txt
    #   coreapi
ruamel-yaml==0.17.21
    # via
    #   -r requirements.txt
    #   drf-yasg
ruamel-yaml-clib==0.2.7
    # via
    #   -r requirements.txt
    #   ruamel-yaml
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via
    #   -r requirements.txt
    #   django
    #   django-debug-toolbar
swagger-spec-validator==3.0.3
    # via
    #   -r requirements.txt
    #   drf-yasg
tomli==2.0.1
    # via black
typing-extensions==4.4.0
    # via
    #   -r requirements.txt
    #   black
    #   swagger-spec-validator
uritemplate==4.1.1
    # via
    #   -r requirements.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.14
    # via
    #   -r requirements.txt
    #   requests
uwsgi==2.0.21
    # via -r requirements.txt
zipp==3.12.0
    # via
    #   -r requirements.txt
    #   importlib-metadata
